server/requirements.txt
uvicorn==0.27.1
fastapi==0.109.2
pydantic==2.6.2
pycryptodome==3.20.0